add controllers, routes and seeds
