+++ config/routes.rb
@@ -1,14 +1,18 @@
 Rails.application.routes.draw do
   
-  resources :users do 
-    resources :reservations
+  namespace :Api do
+    namespace :V1 do
+      resources :users do 
+        resources :reservations
+      end
+    end
   end
 
-  resources :events do 
-    resources :reservations
-  end
-  # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
-
-  # Defines the root path route ("/")
-  # root "articles#index"
+  # namespace :Api do
+  #   namespace :V1 do
+  #     resources :events do 
+  #       resources :reservations
+  #     end
+  #   end
+  # end
 end
